requirements.txt
domdf-python-tools>=3.10.0
github3-utils>=0.7.1
github3.py>=4.0.1
packaging>=25.0
python-dotenv>=1.0.1
repo-helper-github>=0.9.0
ruamel.yaml>=0.18.16
southwark>=1.0.0
